feat(dsh): promote deepseek-harness adapter to first-class goal-mode subpackage
Move the DeepSeek Harness Turn host adapter from the loose
scripts/dsh_turn_host_adapter.py launcher into loopx/dsh_goal_mode/,
matching the pi/opencode/kunluncode goal-mode packaging pattern:

- loopx/dsh_goal_mode/turn_host_adapter.py is the canonical implementation;
  run it with python -m loopx.dsh_goal_mode.
- scripts/dsh_turn_host_adapter.py stays as a backward-compatible launcher
  re-exporting the subpackage, so existing commands, examples, and docs keep
  working.
- Point the agent-type catalog, host-loop activation packet, onboarding
  instruction, and bootstrap command-pack description at the new module while
  keeping the legacy path documented.
- Add README and package-data entry, cover the subpackage and the compat
  launcher with tests (hermetic fake dsh runner), and update README/docs.
