+++ docs/reference/protocols/loopx-turn-v0.md
@@ -87,7 +87,9 @@ A reference headless adapter for TraeX (`traex exec`) lives at
 Turn envelope and writes one typed result, leaving goal/todo authority and
 validation to LoopX.
 
-A DeepSeek Harness adapter lives at `scripts/dsh_turn_host_adapter.py`; it uses
+A DeepSeek Harness adapter lives in the `loopx.dsh_goal_mode` subpackage
+(run with `python -m loopx.dsh_goal_mode`; the legacy
+`scripts/dsh_turn_host_adapter.py` launcher still works); it uses
 the optional `deepseek-harness-sdk` Python client to run one bounded dsh session
 and parses the final assistant JSON message into the same typed Turn result.
 See [DeepSeek Harness connector](../../integrations/deepseek-harness-connector.md).
