+++ loopx/dsh_goal_mode/__main__.py
@@ -1,0 +1,8 @@
+"""Run the DeepSeek Harness host adapter with ``python -m loopx.dsh_goal_mode``."""
+
+from __future__ import annotations
+
+from .turn_host_adapter import main
+
+if __name__ == "__main__":
+    raise SystemExit(main())
