+++ pyproject.toml
@@ -53,6 +53,7 @@ include = ["loopx*"]
 "loopx.extensions.lark" = ["extension.toml", "README.md", "docs/*.md"]
 "loopx.extensions.openviking_periodic_report" = ["extension.toml"]
 "loopx.extensions.openviking_semantic_preference" = ["extension.toml"]
+"loopx.dsh_goal_mode" = ["README.md"]
 "loopx.opencode_goal_mode" = ["*.js", "*.mjs", "README.md"]
 "loopx.opencode2_goal_mode" = ["*.mjs", "README.md"]
 "loopx.pi_goal_mode" = ["*.ts", "*.mjs", "README.md"]
